goquant_trade_simulator/requirements.txt
numpy
scikit-learn
PyQt5
websocket-client
